pyproject.toml
[tool.poetry]
name = "langchain-chatbot"
version = "0.1.0"
description = ""
authors = ["Sai Ram Madala"]
license = "MIT"
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.10"
langchain = "^0.1.11"
langchain-openai = "^0.0.8"
python-dotenv = "^1.0.1"
pinecone-client = "2.2.1"
colorama = "^0.4.6"
streamlit = "^1.32.0"
watchdog = "^4.0.0"
python-multipart = "^0.0.9"
pymupdf = "^1.23.26"
fastapi = "^0.110.0"
uvicorn = "^0.28.0"
langchain-anthropic = "^0.1.4"
docx2txt = "^0.8"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
